pyproject.toml
[project]
name = "tengu"
version = "0.1.0"
description = "A Python package for processing and analyzing Visium HD spatial transcriptomics data."
authors = [
    { name = "Visanu Wanchai", email = "vwanchai@uams.edu" },
]
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "scanpy",
    "anndata",
    "pandas",
    "matplotlib",
    "stardist",
    "tifffile",
    "seaborn",
    "scikit-image",
    "scipy",
    "numpy",
    "Pillow",
    "opencv-python",
    "h5py",
]

[project.urls]
Homepage = "https://github.com/visanuwan/TENGU"